Struct cargopants::Client

pub struct Client {
    // some fields omitted
}

Entry point for accessing crates.io

Methods

impl Client

fn new() -> Client

Create a new Client interface for crates.io

fn host(addr: &str) -> Client

Create a new Client interface for a given host

fn token(self, auth: &str) -> Client

Authenticate requests with an auth token. You can retrieve a cargo auth token from crates.io

examples

let client = cargopants::Client::new().token("token");

fn krate<'a>(&'a mut self, name: &'a str) -> Krate

Provides access to crate-specific resources

fn find(&mut self, query: &str) -> Result<Vec<Crate>>

Issues a request to find a crate by name

fn publish(&mut self, krate: &NewCrate, tarball: &Path) -> Result<()>

Publishes a tar'd crate file to crate server