Struct cargopants::Krate

pub struct Krate<'a, 'b> {
    // some fields omitted
}

Client interface for a given crate

Methods

impl<'a, 'b> Krate<'a, 'b>

fn new(client: &'a mut Client, name: &'b str) -> Krate<'a, 'b>

Provides a crate specific view of information

fn downloads(self) -> Result<Vec<Download>>

Request download information for a create

fn follow(self) -> Result<()>

Follow a crate

fn unfollow(self) -> Result<()>

Unfollow a crate

fn following(self) -> Result<bool>

Request indication of whether the current authentication credentials follows this crate

fn get(self) -> Result<Crate>

Get the base set of information associated for a crate

fn owners(self) -> Result<Vec<User>>

Requests a vector of owners for a crate

fn add_owners(self, owners: &[&str]) -> Result<()>

Adds owner to for a crate

fn remove_owners(self, owners: &[&str]) -> Result<()>

Remove owners from a crate

fn reverse_dependencies(&mut self) -> Result<Vec<Dependency>>

Fetches references to crates that depend on this crate

fn version<'c>(&'c mut self, version: &'c str) -> KrateVersion

Provides access to crate version specific resources

fn versions(self) -> Result<Vec<Version>>

Requests all versions associated with a given create